requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
albumentations>=1.0.0
Pillow>=8.3.0
gradio>=6.1.0